调整点位运动和回馈的计算

+++ GTSSimpleController/Mgr/MotionMgr.h
@@ -32,12 +32,12 @@ private:
 private:
     // 检查规划器
     bool checkProfile(short profile) const;
-    // 点位运动脉冲单位
-    bool singleTrapMotion(short profile, long stepSize, double acc, double dec, int smoothTime, double vel);
+    // 点位运动
+    bool singleTrapMotion(short profile, double stepSize, double acc, double dec, int smoothTime, double vel);
     // 等待运动
     void waitMotionDone(short profile);
     // 启动点位运动
-    bool startTrapMotion(short profile, long stepSize, const stuTrapParam& trapParam);
+    bool startTrapMotion(short profile, double stepSize, const stuTrapParam& trapParam);
 public:
     explicit MotionMgr(CTotalMgr* totalMgr, QObject* parent = nullptr);
     ~MotionMgr();
